+++ control.go
@@ -148,7 +148,7 @@ func hostInfo(addr string, defaultPort int) ([]*HostInfo, error) {
 
 	// Check if host is a literal IP address
 	if ip := net.ParseIP(host); ip != nil {
-		h, err := NewHostInfo(ip, port)
+		h, err := NewHostInfoFromAddrPort(ip, port)
 		if err != nil {
 			return nil, err
 		}
@@ -178,7 +178,7 @@ func hostInfo(addr string, defaultPort int) ([]*HostInfo, error) {
 	}
 
 	for _, ip := range ips {
-		h, err := NewHostInfo(ip, port)
+		h, err := NewHostInfoFromAddrPort(ip, port)
 		if err != nil {
 			return nil, err
 		}